refactor: split code graph refresh pipeline
